# simplify the `query` implementation