This code is drastically simpler than the original implementation, but also much slower:

Kept here as a reference and for potential future work that might be both simpler and faster.

```
## before any simplification

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  134.635 μs …   6.296 ms  ┊ GC (min … max): 0.00% … 95.01%
 Time  (median):     142.902 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   160.930 μs ± 240.887 μs  ┊ GC (mean ± σ):  8.04% ±  5.30%

    ▂▇██▅▂
  ▃▅███████▇▆▅▅▅▅▅▅▅▄▄▄▃▃▃▃▃▃▃▃▃▃▃▂▂▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂ ▃
  135 μs           Histogram: frequency by time          205 μs <

 Memory estimate: 214.33 KiB, allocs estimate: 2649.

## after first round of simplification

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  171.124 μs …  10.533 ms  ┊ GC (min … max): 0.00% … 96.97%
 Time  (median):     182.391 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   200.198 μs ± 291.337 μs  ┊ GC (mean ± σ):  6.32% ±  4.33%

    ▁▆███▇▃▂
  ▁▃██████████▇▇▆▅▄▄▄▃▃▃▂▂▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  171 μs           Histogram: frequency by time          256 μs <

 Memory estimate: 226.64 KiB, allocs estimate: 2848.

## after second round of simplification

BenchmarkTools.Trial: 10000 samples with 1 evaluation.
 Range (min … max):  331.740 μs …   6.718 ms  ┊ GC (min … max): 0.00% … 91.94%
 Time  (median):     344.995 μs               ┊ GC (median):    0.00%
 Time  (mean ± σ):   368.845 μs ± 261.286 μs  ┊ GC (mean ± σ):  5.52% ±  7.07%

     ▅██▆▃▂
  ▂▃▇███████▆▆▅▄▃▃▃▃▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁▂▂▂▂▁▁▁▁▁▂▂▁▁▁▁▁▁▂▂▂▁▂ ▃
  332 μs           Histogram: frequency by time          458 μs <

 Memory estimate: 436.03 KiB, allocs estimate: 8666.
```